requirements.txt
openai>=1.30.0
transformers>=4.40.0
torch>=2.2.0
accelerate>=0.28.0
pydantic>=2.6.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
datasets>=2.18.0
scikit-learn>=1.3.0
sympy>=1.12

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
